Ver 0.0.3 - architecture refactor
Start of refactor of architecture which is moving towards Engine holding atomic references to various modules which require interaction. The previous engine.rs is migrated to usermode_api.rs. This will be decommissioned in the near future and any state management will be moved into modules; or if that's not best suited, managed in engine.rs.